Add `HeightRound` for round representation (#4)
This implementation aims to be closer to Tendermint. There, the round
implementation isn't generic and is always a tuple (height, round).

This provides the ability to perform concrete implementations for some
parts of the consensus, such as the leader election per round. In
Tendermint, it is performed as a round-robin of the leaders.

The library will run a round-robin per height over a list of sorted
validators. The list of sorted validators must be provided by the
implementor, and the sorting strategy is considered an implementation
detail.

It can and is expected that the implementor optimizes and cache the
sorted validators so this list can be sorted efficiently. This will be
an associated type of the network trait, so its internal logic can
assume the sorting is solved by an efficient algorithm.

The implementor can overwrite any of these implementations with more
efficient or convenient implementations, when applicable.