Add Stake and StakePool structs (#18)
The stake structure will be needed to address #17. We will define the
voting power from the total staked amount for a given height, and
calculate bft out of that.

Two new types were introduced: Stake and StakePool

Stake is a minimalistic representation of a staked value with an
associated consensus key used to verify the round signatures.

StakePool is a mapping from an external validator identifier, constant
per identity, to a set of stakes mapped by block height ranges.